add GenericRecipeBookComponent.mouseClicked
